Move attribute validation from `rustc_parse` to `rustc_attr_parsing`
